+++ android/src/main/kotlin/dev/steenbakker/mobile_scanner/MobileScannerPlugin.kt
@@ -1,50 +1,46 @@
 package dev.steenbakker.mobile_scanner
 
-import androidx.annotation.NonNull
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
-import io.flutter.plugin.common.EventChannel
-import io.flutter.plugin.common.MethodChannel
 
 /** MobileScannerPlugin */
 class MobileScannerPlugin : FlutterPlugin, ActivityAware {
-    private var flutter: FlutterPlugin.FlutterPluginBinding? = null
-    private var activity: ActivityPluginBinding? = null
-    private var handler: MobileScanner? = null
-    private var method: MethodChannel? = null
-    private var event: EventChannel? = null
+    private var activityPluginBinding: ActivityPluginBinding? = null
+    private var flutterPluginBinding: FlutterPlugin.FlutterPluginBinding? = null
+    private var methodCallHandler: MobileScannerHandler? = null
 
-    override fun onAttachedToEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-        this.flutter = binding
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        this.flutterPluginBinding = binding
     }
 
-    override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-        this.flutter = null
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        this.flutterPluginBinding = null
     }
 
-    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
-        activity = binding
-        handler = MobileScanner(activity!!.activity, flutter!!.textureRegistry)
-        method = MethodChannel(flutter!!.binaryMessenger, "dev.steenbakker.mobile_scanner/scanner/method")
-        event = EventChannel(flutter!!.binaryMessenger, "dev.steenbakker.mobile_scanner/scanner/event")
-        method!!.setMethodCallHandler(handler)
-        event!!.setStreamHandler(handler)
-        activity!!.addRequestPermissionsResultListener(handler!!)
+    override fun onAttachedToActivity(activityPluginBinding: ActivityPluginBinding) {
+        val binaryMessenger = this.flutterPluginBinding!!.binaryMessenger
+
+        methodCallHandler = MobileScannerHandler(
+            activityPluginBinding.activity,
+            BarcodeHandler(binaryMessenger),
+            binaryMessenger,
+            MobileScannerPermissions(),
+            activityPluginBinding::addRequestPermissionsResultListener,
+            this.flutterPluginBinding!!.textureRegistry,
+        )
+
+        this.activityPluginBinding = activityPluginBinding
+    }
+
+    override fun onDetachedFromActivity() {
+        methodCallHandler?.dispose(this.activityPluginBinding!!)
+        methodCallHandler = null
+        activityPluginBinding = null
     }
 
     override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
         onAttachedToActivity(binding)
-    }
-
-    override fun onDetachedFromActivity() {
-        activity!!.removeRequestPermissionsResultListener(handler!!)
-        event!!.setStreamHandler(null)
-        method!!.setMethodCallHandler(null)
-        event = null
-        method = null
-        handler = null
-        activity = null
     }
 
     override fun onDetachedFromActivityForConfigChanges() {