+++ macos/Classes/MobileScannerPlugin.swift
@@ -1,6 +1,7 @@
 import AVFoundation
 import FlutterMacOS
 import Vision
+import AppKit
 
 public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler, FlutterTexture, AVCaptureVideoDataOutputSampleBufferDelegate {
     
@@ -20,6 +21,11 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
     
     // Image to be sent to the texture
     var latestBuffer: CVImageBuffer!
+
+    // optional window to limit scan search
+    var scanWindow: CGRect?
+    
+    var detectionSpeed: DetectionSpeed = DetectionSpeed.noDuplicates
     
     
 //    var analyzeMode: Int = 0
@@ -57,6 +63,8 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
 //            switchAnalyzeMode(call, result)
         case "stop":
             stop(result)
+        case "updateScanWindow":
+            updateScanWindow(call)
         default:
             result(FlutterMethodNotImplemented)
         }
@@ -87,51 +95,56 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
     
     // Gets called when a new image is added to the buffer
     public func captureOutput(_ output: AVCaptureOutput, didOutput sampleBuffer: CMSampleBuffer, from connection: AVCaptureConnection) {
-        i = i + 1;
-        
-        latestBuffer = CMSampleBufferGetImageBuffer(sampleBuffer)
+        // Ignore invalid textureId
+        if textureId == nil {
+            return
+        }
+        guard let imageBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else {
+            print("Failed to get image buffer from sample buffer.")
+            return
+        }
+        latestBuffer = imageBuffer
         registry.textureFrameAvailable(textureId)
-
-//        switch analyzeMode {
-//        case 1: // barcode
+        
+        if ((detectionSpeed == DetectionSpeed.normal || detectionSpeed == DetectionSpeed.noDuplicates) && i > 10 || detectionSpeed == DetectionSpeed.unrestricted) {
+            i = 0
+            let imageRequestHandler = VNImageRequestHandler(
+                cvPixelBuffer: latestBuffer,
+                orientation: .right)
             
-            // Limit the analyzer because the texture output will freeze otherwise
-            if i / 10 == 1 {
-                i = 0
-            } else {
-                return
-            }
-                let imageRequestHandler = VNImageRequestHandler(
-                    cvPixelBuffer: latestBuffer,
-                    orientation: .right)
+                do {
+                  try imageRequestHandler.perform([VNDetectBarcodesRequest { (request, error) in
+                      if error == nil {
+                          if let results = request.results as? [VNBarcodeObservation] {
+                                for barcode in results {
+                                    if self.scanWindow != nil {
+                                        let match = self.isbarCodeInScanWindow(self.scanWindow!, barcode, self.latestBuffer)
+                                        if (!match) {
+                                            continue
+                                        }
+                                    }
 
-            do {
-              try imageRequestHandler.perform([VNDetectBarcodesRequest { (request, error) in
-                  if error == nil {
-                      if let results = request.results as? [VNBarcodeObservation] {
-                                  for barcode in results {
-                                      let barcodeType = String(barcode.symbology.rawValue).replacingOccurrences(of: "VNBarcodeSymbology", with: "")
-                                      let event: [String: Any?] = ["name": "barcodeMac", "data" : ["payload": barcode.payloadStringValue, "symbology": barcodeType]]
-                                      self.sink?(event)
+                                    let barcodeType = String(barcode.symbology.rawValue).replacingOccurrences(of: "VNBarcodeSymbology", with: "")
+                                    let event: [String: Any?] = ["name": "barcodeMac", "data" : ["payload": barcode.payloadStringValue, "symbology": barcodeType]]
+                                    self.sink?(event)
 
-  //                                    if barcodeType == "QR" {
-  //                                        let image = CIImage(image: source)
-  //                                        image?.cropping(to: barcode.boundingBox)
-  //                                        self.qrCodeDescriptor(qrCode: barcode, qrCodeImage: image!)
-  //                                    }
-                                  }
+        //                                    if barcodeType == "QR" {
+        //                                        let image = CIImage(image: source)
+        //                                        image?.cropping(to: barcode.boundingBox)
+        //                                        self.qrCodeDescriptor(qrCode: barcode, qrCodeImage: image!)
+        //                                    }
+                                      }
+                          }
+                      } else {
+                          print(error!.localizedDescription)
                       }
-                  } else {
-                      print(error!.localizedDescription)
-                  }
-              }])
-            } catch {
-              print(error)
-            }
-
-//        default: // none
-//            break
-//        }
+                  }])
+                } catch {
+                  print(error)
+                }
+        } else {
+            i+=1
+        }
     }
     
     func checkPermission(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
@@ -158,6 +171,38 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
         }
     }
 
+    func updateScanWindow(_ call: FlutterMethodCall) {
+        let argReader = MapArgumentReader(call.arguments as? [String: Any])
+        let scanWindowData: Array? = argReader.floatArray(key: "rect")
+
+        if (scanWindowData == nil) {
+            return 
+        }
+
+        let minX = scanWindowData![0] 
+        let minY = scanWindowData![1]
+
+        let width = scanWindowData![2]  - minX
+        let height = scanWindowData![3] - minY
+
+        scanWindow = CGRect(x: minX, y: minY, width: width, height: height)
+    }
+
+    func isbarCodeInScanWindow(_ scanWindow: CGRect, _ barcode: VNBarcodeObservation, _ inputImage: CVImageBuffer) -> Bool {
+       let size = CVImageBufferGetEncodedSize(inputImage)
+
+        let imageWidth = size.width;
+        let imageHeight = size.height;
+
+        let minX = scanWindow.minX * imageWidth
+        let minY = scanWindow.minY * imageHeight
+        let width = scanWindow.width * imageWidth
+        let height = scanWindow.height * imageHeight
+
+        let scaledScanWindow = CGRect(x: minX, y: minY, width: width, height: height)
+        return scaledScanWindow.contains(barcode.boundingBox)
+    }
+
     func start(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
         if (device != nil) {
             result(FlutterError(code: "MobileScanner",
@@ -174,6 +219,9 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
 //        let ratio: Int = argReader.int(key: "ratio")
         let torch: Bool = argReader.bool(key: "torch") ?? false
         let facing: Int = argReader.int(key: "facing") ?? 1
+        let speed: Int = (call.arguments as! Dictionary<String, Any?>)["speed"] as? Int ?? 0
+        
+        detectionSpeed = DetectionSpeed(rawValue: speed)!
 
         // Set the camera to use
         position = facing == 0 ? AVCaptureDevice.Position.front : .back
@@ -238,9 +286,7 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
     
     func toggleTorch(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
         if (device == nil) {
-            result(FlutterError(code: "MobileScanner",
-                                    message: "Called toggleTorch() while stopped!",
-                                    details: nil))
+            result(nil)
             return
         }
         do {
@@ -260,9 +306,8 @@ public class MobileScannerPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
 
     func stop(_ result: FlutterResult) {
         if (device == nil) {
-            result(FlutterError(code: "MobileScanner",
-                                    message: "Called stop() while already stopped!",
-                                    details: nil))
+            result(nil)
+
             return
         }
         captureSession.stopRunning()
@@ -320,6 +365,10 @@ class MapArgumentReader {
 
   func stringArray(key: String) -> [String]? {
     return args?[key] as? [String]
+  }
+
+  func floatArray(key: String) -> [CGFloat]? {
+    return args?[key] as? [CGFloat]
   }
   
 }